Adapted imdb example from PDL course. [work in progress]
Bit less evolved set up for the dataset and no concatenation of training samples.

Runs 5000 steps with batch size 16 on single GPU in ~10 minutes.
Tried to adapt to multigpu as well, but not getting it to run properly just yet. Refuses to run without torchrun. With torchrun it runs but takes the same amount of time, so I suspect it just spawn two independent
processes that don't talk to each other. Needs fixing.

Uses CSC pytorch module so far, will need to adjust.